pyproject.toml
[tool.poetry]
name = "drf-course"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
pillow = "^10.3.0"
config = "^0.5.1"
loadenv = "^0.1.1"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
drf-spectacular = "^0.27.2"
setuptools = "^70.0.0"
drf-spectacular-sidecar = "^2024.5.1"
django-filter = "^24.2"
django-redis = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
